Update end trajectory test to be timer-based
This switches how the joint trajectory action detects the end of a
trajectory. Previously, we would wait for the robot to be in-motion
then exit being in-motion. However, this caused stalling issues for
short trajectories. Now, we wait until the time expected to complete
the trajectory has passed and the robot is not in motion.